Added Die Chart to README.txt
I was testing out the Queller java program and realized I didn't know the code for the "Will of the West/Eye" dice result. I looked in the README but didn't find anything. I dug through the code and found the die chart, but the F character (for Faction) didn't work either. I had to comb through the ConstantUtil.java file to find that P was the character for the Event die face, and E was the character for the Eye die face.

I suggest adding in the die chart to the README as a reference for those unfamiliar with the Queller java program syntax. Thank you for your work on the bot, it's really fun to play against when other opponents are unavailable!